pyproject.toml
[tool.poetry]
name = "rag-beyond-the-basics"
version = "0.1.0"
description = ""
authors = ["Adam Hosman <github@adamsinbox.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
python-dotenv = "1.0.1"
openai = "1.53.0"
langchain-openai = "0.2.5"
langgraph = "0.2.44"
bs4 = "0.0.2"
langchain = "0.3.7"
langchain-chroma = "0.1.4"
langchain-community = "0.3.5"
langchain-text-splitters = "0.3.2"
deepeval = "1.4.7"
click = "8.1.7"
rank-bm25 = "0.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
ws = "src.main.util.cli:ws"
